+++ src/platform_impl/ios/uikit/mod.rs
@@ -12,7 +12,6 @@ mod event;
 mod responder;
 mod screen;
 mod screen_mode;
-mod status_bar;
 mod touch;
 mod trait_collection;
 mod view;
@@ -26,7 +25,6 @@ pub(crate) use self::event::UIEvent;
 pub(crate) use self::responder::UIResponder;
 pub(crate) use self::screen::{UIScreen, UIScreenOverscanCompensation};
 pub(crate) use self::screen_mode::UIScreenMode;
-pub(crate) use self::status_bar::UIStatusBarStyle;
 pub(crate) use self::touch::{UITouch, UITouchPhase, UITouchType};
 pub(crate) use self::trait_collection::{UIForceTouchCapability, UITraitCollection};
 #[allow(unused_imports)]
